+++ db/migrate/20110301204015_create_area_counts.rb
@@ -2,7 +2,7 @@ class CreateAreaCounts < ActiveRecord::Migration
   def self.up
     create_table :area_counts do |t|
       t.integer :calibration_level_id
-      t.integer :area
+      t.float :area
       t.string :analyte
 
       t.timestamps
